+++ src/P8PGCLogger.h
@@ -48,7 +48,7 @@ namespace NVDSim
     public:
 	P8PGCLogger();
 	
-	void save(uint64_t cycle, uint epoch);
+	void save(uint64_t cycle, uint64_t epoch);
 	void print(uint64_t cycle);
 
 	void update();
@@ -59,7 +59,7 @@ namespace NVDSim
 	//Writing correct object oriented code up in this piece, what now?
 	std::vector<std::vector<double>> getEnergyData(void);
 
-	void save_epoch(uint64_t cycle, uint epoch);
+	void save_epoch(uint64_t cycle, uint64_t epoch);
 
 	// State
 	// Power Stuff
