+++ src/Plane.h
@@ -55,7 +55,7 @@ namespace NVDSim{
 			void dataGone(void);
 		private:
 			ChannelPacket *dataReg, *cacheReg;
-			std::unordered_map<uint, Block> blocks;
+			std::unordered_map<uint64_t, Block> blocks;
 	};
 }
 #endif
